refactor: collapse runtime ecstore thin bridges
